final_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2024.12.0
google-auth==2.43.0
google-genai==1.53.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-mongodb==0.9.0
langchain-neo4j==0.6.0
langchain-openai==1.1.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-mongodb==0.3.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.4.58
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==1.0.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.11.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.9.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
portalocker==3.2.0
posthog==7.0.1
propcache==0.4.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
pymongo-search-utils==0.1.0
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
rank-bm25==0.2.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rq==2.6.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.47.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.4
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer==0.20.0
typer-slim==0.20.0
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
